YAY-79 fix: change loadImage, dataURL, ref: transfered avatar logic from page
