+++ src/entities/settings/ui/modals/AddAvatarModal/index.ts
@@ -1,2 +1,3 @@
 export * from './AddAvatarModal'
 export * from './AvatarCropper'
+export * from './AvatarUploader'
